+++ src/client/src/renderer/gui/panels/singleplayer_panel/SingleplayerPanel.cpp
@@ -684,7 +684,7 @@ namespace onion::voxel
 			}
 			else
 			{
-				m_SelectedWorldIndex = i;
+				m_SelectedWorldIndex = static_cast<int>(i);
 			}
 		}
 	}
